Change 'Url encode' to 'Percent-encode'
Thank you for this helpful tool.  

At first, I found it a bit confusing that it had an option to URL-encode paths, as URLs and paths are based on different standards.  I think it will help to change this label to 'Percent-encode' as that is the term preferred by RFC 3986 as well as Wikipedia:  

rfc3986  
https://datatracker.ietf.org/doc/html/rfc3986#page-12

Binary-to-text encoding - Wikipedia  
https://en.wikipedia.org/wiki/Binary-to-text_encoding

Percent-encoding - Wikipedia  
https://en.wikipedia.org/wiki/Percent-encoding